Fixes bug in _cleanUp method of DVRconverter
In the DVRconverter class, had some checks in the _cleanUp
method, however did not check that the inputs were valid; i.e.,
not None. Added check for valid before running any os. commands.

v0.10.3

+++ video_utils/plex/DVRconverter.py
@@ -86,7 +86,7 @@ class DVRconverter(comremove, videoconverter):
     Method to delete arbitary number of files, catching exceptions
     '''
     for arg in args:
-      if os.path.isfile(arg):
+      if arg and os.path.isfile(arg):
         try:
           os.remove(arg)
         except Exception as err:
